refactor: 도메인 로직과 UI 관련 로직 폴더 구조 변경
