+++ src/domain/constants/Conditions.js
