+++ src/domain/constants/Errors.js
